feat(dedicated): rename subrange and creat constant file
