+++ packages/manager/modules/vrack/src/vrack-associated-services/ipv6/ipv6.constant.js
@@ -1,0 +1,36 @@
+export const SLAAC_GUIDES_LINK = {
+  DE:
+    'https://help.ovhcloud.com/csm/de-dedicated-servers-configure-an-ipv6-in-a-vrack?id=kb_article_view&sysparm_article=KB0062832',
+  ASIA:
+    'https://help.ovhcloud.com/csm/asia-dedicated-servers-configure-an-ipv6-in-a-vrack?id=kb_article_view&sysparm_article=KB0062833',
+  AU:
+    'https://help.ovhcloud.com/csm/en-au-dedicated-servers-configure-an-ipv6-in-a-vrack?id=kb_article_view&sysparm_article=KB0062836',
+  CA:
+    'https://help.ovhcloud.com/csm/en-ca-dedicated-servers-configure-an-ipv6-in-a-vrack?id=kb_article_view&sysparm_article=KB0062827',
+  GB:
+    'https://help.ovhcloud.com/csm/en-gb-dedicated-servers-configure-an-ipv6-in-a-vrack?id=kb_article_view&sysparm_article=KB0062828',
+  IE:
+    'https://help.ovhcloud.com/csm/en-ie-dedicated-servers-configure-an-ipv6-in-a-vrack?id=kb_article_view&sysparm_article=KB0062826',
+  SG:
+    'https://help.ovhcloud.com/csm/en-sg-dedicated-servers-configure-an-ipv6-in-a-vrack?id=kb_article_view&sysparm_article=KB0062829',
+  EN:
+    'https://help.ovhcloud.com/csm/en-dedicated-servers-configure-an-ipv6-in-a-vrack?id=kb_article_view&sysparm_article=KB0062838',
+  ES:
+    'https://help.ovhcloud.com/csm/es-es-dedicated-servers-configure-an-ipv6-in-a-vrack?id=kb_article_view&sysparm_article=KB0062830',
+  WS:
+    'https://help.ovhcloud.com/csm/es-dedicated-servers-configure-an-ipv6-in-a-vrack?id=kb_article_view&sysparm_article=KB0062835',
+  QC:
+    'https://help.ovhcloud.com/csm/fr-ca-dedicated-servers-configure-an-ipv6-in-a-vrack?id=kb_article_view&sysparm_article=KB0062834',
+  FR:
+    'https://help.ovhcloud.com/csm/fr-dedicated-servers-configure-an-ipv6-in-a-vrack?id=kb_article_view&sysparm_article=KB0062825',
+  IT:
+    'https://help.ovhcloud.com/csm/it-dedicated-servers-configure-an-ipv6-in-a-vrack?id=kb_article_view&sysparm_article=KB0062839',
+  PL:
+    'https://help.ovhcloud.com/csm/pl-dedicated-servers-configure-an-ipv6-in-a-vrack?id=kb_article_view&sysparm_article=KB0062837',
+  PT:
+    'https://help.ovhcloud.com/csm/pt-dedicated-servers-configure-an-ipv6-in-a-vrack?id=kb_article_view&sysparm_article=KB0062831',
+  DEFAULT:
+    'https://help.ovhcloud.com/csm/en-ie-dedicated-servers-configure-an-ipv6-in-a-vrack?id=kb_article_view&sysparm_article=KB0062826',
+};
+
+export const SLAAC_LABEL = 'SLAAC';
