+++ packages/manager/modules/vrack/src/vrack-associated-services/ipv6/ipv6.less
@@ -9,7 +9,7 @@
     }
 
     &__subnet,
-    &__subrang {
+    &__subrange {
       list-style: none;
       padding: 0.625rem 0.9375rem;
       border-width: 1px;
